Add files via upload
Adding updated version of the GUI. It is mainly a proof of concept. It shows that a login, a classroom creation, and tkinter "Labels" can be edited since the labels are objects and their references are saved in an array. This is mainly the framework and it will have to be edited to completely be implemented with the data base